+++ pxr/usd/lib/sdf/pathNode.cpp
@@ -326,7 +326,7 @@ void Sdf_PathNode::GetPrefixes(SdfPathVector *prefixes, bool includeRoot) const
     size_t nElems = _elementCount + (includeRoot ? 1 : 0);
     prefixes->resize(nElems);
     Sdf_PathNodeConstRefPtr n(this);
-    for (int i = nElems-1; i >= 0; --i) {
+    for (int i = static_cast<int>(nElems)-1; i >= 0; --i) {
         (*prefixes)[i] = SdfPath(n); 
         n = n->_parent;
     }
@@ -650,7 +650,7 @@ void Sdf_DumpPathStats()
 
     printf("------------------------------------------------");
     printf("-- By Length\n");
-    int totalLen = 0;
+    size_t totalLen = 0;
     for (size_t i=0; i < stats.lengthTable.size();++i) {
         printf("\tnum nodes with %3zu components : %i\n",
                i, stats.lengthTable[i] );
@@ -666,7 +666,7 @@ void Sdf_DumpPathStats()
                i, stats.numChildrenTable[i] );
     }
 
-    int numChildren = 0;
+    size_t numChildren = 0;
     for (size_t i=1; i < stats.numChildrenTable.size(); ++i) {
         numChildren += i*stats.numChildrenTable[i];
     }
